% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_utils.R
\name{alias2df}
\alias{alias2df}
\title{Create a matrix of alias}
\usage{
alias2df(x)
}
\arguments{
\item{x}{A raw alias output as given by CRAN (and Bioconductor).}
}
\value{
A matrix with Package, Source (The name of the file), and the Target (or Alias)
}
\description{
Joins the matrices of each file
}
\keyword{internal}
