% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_alias.R
\name{base_alias}
\alias{base_alias}
\title{Base R's alias}
\usage{
base_alias(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with three columns: Package, Source and Target.
\code{NA} if not able to collect the data from CRAN.
}
\description{
Retrieve alias available on R.
}
\examples{
\donttest{
ba <- base_alias()
head(ba)
}
}
\seealso{
The raw source of the data is: \code{\link[tools:base_aliases_db]{base_aliases_db()}}.

Other alias:
\code{\link[=alias]{alias()}},
\code{\link[=cran_alias]{cran_alias()}}
}
\concept{alias}
