% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_xrefs.R
\name{base_links}
\alias{base_links}
\title{Base R's links}
\usage{
base_links(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with the links on R's files.
It has 4 columns: Package, Anchor, Target and Source.
\code{NA} if not able to collect the data from CRAN.
}
\description{
Retrieve links on R documentation files.
}
\examples{
\donttest{
bl <- base_links()
head(bl)
}
}
\seealso{
The raw source of the data is: \code{\link[tools:base_rdxrefs_db]{base_rdxrefs_db()}}.

Other links from R:
\code{\link[=base_pages_links]{base_pages_links()}},
\code{\link[=base_pkges_links]{base_pkges_links()}},
\code{\link[=base_targets_links]{base_targets_links()}}
}
\concept{links from R}
