% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_xrefs.R
\name{base_pages_links}
\alias{base_pages_links}
\title{Links between help pages by page}
\usage{
base_pages_links(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with 6 columns: from_pkg, from_Rd, to_pkg, to_Rd, n (Number of links).
\code{NA} if not able to collect the data from CRAN.
}
\description{
Explore the relationship between base R packages and other help pages.
If the target help page is ambiguous it is omitted.
}
\examples{
\donttest{
bpl <- base_pages_links()
head(bpl)
}
}
\seealso{
Other links from R:
\code{\link[=base_links]{base_links()}},
\code{\link[=base_pkges_links]{base_pkges_links()}},
\code{\link[=base_targets_links]{base_targets_links()}}
}
\concept{links from R}
