% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_xrefs.R
\name{base_pkges_links}
\alias{base_pkges_links}
\title{Links between help pages by package}
\usage{
base_pkges_links(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with 6 columns: from_pkg, to_pkg, n (Number of links).
\code{NA} if not able to collect the data from CRAN.
}
\description{
Explore the relationship between base R packages and other packages.
If the target package is ambiguous it is omitted.
}
\examples{
\donttest{
bpkl <- base_pkges_links()
head(bpkl)
}
}
\seealso{
Other links from R:
\code{\link[=base_links]{base_links()}},
\code{\link[=base_pages_links]{base_pages_links()}},
\code{\link[=base_targets_links]{base_targets_links()}}
}
\concept{links from R}
