% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_xrefs.R
\name{base_targets_links}
\alias{base_targets_links}
\title{Links between help pages by target}
\usage{
base_targets_links(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with 6 columns: from_pkg, from_Rd, to_pkg, to_target, to_Rd, n (Number of links).
\code{NA} if not able to collect the data from CRAN.
}
\description{
Explore the relationship between base R packages and other help pages by the target they use.
}
\examples{
\donttest{
btl <- base_targets_links()
head(btl)
}
}
\seealso{
Other links from R:
\code{\link[=base_links]{base_links()}},
\code{\link[=base_pages_links]{base_pages_links()}},
\code{\link[=base_pkges_links]{base_pkges_links()}}
}
\concept{links from R}
