% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{clean_cache}
\alias{clean_cache}
\title{Clean cache}
\usage{
clean_cache()
}
\value{
Returns \code{NULL}.
}
\description{
Clean cache to download repository data again.
}
\details{
Cleans the package's environment used for caching the data.
}
\examples{
clean_cache()

}
