% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_comments.R
\name{cran_comments}
\alias{cran_comments}
\title{CRAN comments}
\usage{
cran_comments(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with four columns: package, comment, date and action.
\code{NA} if not able to collect the data from CRAN.
}
\description{
CRAN volunteers document since ~2009 why they archive packages.
This function retrieves the data and prepares it for analysis, classifying
the actions taken by the team per package and date.
}
\details{
The comments are slightly edited: multiple comments for the same action are
joined together so that they can be displayed on a single line.
Actions are inferred from 7 keywords: archived, orphaned, removed, renamed,
replaced, unarchived, unorphaned.
}
\note{
There can be room for improvement: some comments describe two actions, please
let me know if you think this can be improved.
Other actions can be described on multiple comments/lines or out of order.
Compare with the original file in case of doubts.
}
\examples{
\donttest{
cc <- cran_comments()
head(cc)
}
}
\references{
Original file: \url{https://cran.r-project.org/src/contrib/PACKAGES.in}
}
\seealso{
Other meta info from CRAN: 
\code{\link{cran_alias}()},
\code{\link{cran_archive}()},
\code{\link{cran_links}()},
\code{\link{links}()}
}
\concept{meta info from CRAN}
