% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_help.R
\name{cran_help_pages_links_wo_deps}
\alias{cran_help_pages_links_wo_deps}
\title{Links without dependencies}
\usage{
cran_help_pages_links_wo_deps(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame of help pages and links.
\code{NA} if not able to collect the data from CRAN.
}
\description{
On \href{https://cran.r-project.org/doc/manuals/r-devel/R-exts.html#Cross_002dreferences}{WRE section "2.5 Cross-references"} explains that packages shouldn't link
to help pages outside the dependency.
}
\examples{
evmix <- cran_help_pages_links_wo_deps("evmix")
}
\references{
\url{https://cran.r-project.org/doc/manuals/r-devel/R-exts.html#Cross_002dreferences}
}
