% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_history.R
\name{cran_history}
\alias{cran_history}
\title{Check package history on CRAN}
\usage{
cran_history(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with the information to recreate CRAN at any point before today.
\code{NA} if not able to collect the data from CRAN.
}
\description{
It uses available information to provide the most accurate information of
CRAN at any given time.
}
\details{
Several sources are used: CRAN's database to check packages files and versions,
CRAN's volunteers actions for when packages are archived or removed and
CRAN's comments to fill in the gaps.
}
\examples{
cran_history
}
\seealso{
\code{\link[=cran_archive]{cran_archive()}}, \code{\link[=cran_actions]{cran_actions()}}, \code{\link[=cran_comments]{cran_comments()}}.

Other meta info from CRAN: 
\code{\link{cran_alias}()},
\code{\link{cran_archive}()},
\code{\link{cran_comments}()},
\code{\link{cran_links}()}
}
\concept{meta info from CRAN}
