% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_maintainers.R
\name{cran_maintainers}
\alias{cran_maintainers}
\title{Tidy information about maintainers}
\usage{
cran_maintainers()
}
\value{
A data.frame with one row per package and 11 columns.
The package name, Maintainer field, user maintainer manual date, packaged
date, published date, name of maintainer used, email used, direction and domain.
\code{NA} if not able to collect the data from CRAN.
}
\description{
Make more accessible information about maintainers. Extracts and makes
comparable some dates. It also provides the user name used and cleans the
names of the maintainer of extra quotes.
}
\details{
User is what the machine building the package reported. This might indicate
some collaboration, repackaging or simply nothing as it can be hidden/modified.
The name, email and user might help identify similar named people (or confuse between maintainers)
}
\examples{
maintainers <- cran_maintainers()
head(maintainers)
}
\seealso{
The raw source of the data is: \code{\link[tools:CRAN_authors_db]{CRAN_authors_db()}}
}
