% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_xrefs.R
\name{cran_pkges_links}
\alias{cran_pkges_links}
\title{Links between help pages by package}
\usage{
cran_pkges_links(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with 6 columns: from_pkg, to_pkg, n (Number of links).
\code{NA} if not able to collect the data from CRAN.
}
\description{
Explore the relationship between CRAN packages and other packages.
If the target package is ambiguous it is omitted.
}
\examples{
\donttest{
cpkl <- cran_pkges_links()
head(cpkl)
}
}
\seealso{
Other links from CRAN:
\code{\link[=cran_links]{cran_links()}},
\code{\link[=cran_pages_links]{cran_pages_links()}},
\code{\link[=cran_targets_links]{cran_targets_links()}},
\code{\link[=links]{links()}}
}
\concept{links from CRAN}
