% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_xrefs.R
\name{cran_targets_links}
\alias{cran_targets_links}
\title{Links between help pages by target}
\usage{
cran_targets_links(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with 6 columns: from_pkg, from_Rd, to_pkg, to_target, to_Rd, n (Number of links).
}
\description{
Explore the relationship between CRAN packages and other help pages by the target they use.
}
\examples{
\donttest{
ctl <- cran_targets_links("BaseSet")
head(ctl)
}
}
\seealso{
Other links from CRAN: 
\code{\link{cran_links}()},
\code{\link{cran_pages_links}()},
\code{\link{cran_pkges_links}()},
\code{\link{links}()}
}
\concept{links from CRAN}
