% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_version.R
\name{cran_version}
\alias{cran_version}
\title{Install a specific version of a package}
\usage{
cran_version(package, version, ...)
}
\arguments{
\item{package}{Name of the package present on CRAN archive.}

\item{version}{The version number.}

\item{...}{Other arguments passed to install.packages.}
}
\value{
Same as \code{install.packages()}.
}
\description{
Install a package from CRAN of a specific version.
}
\details{
Uses CRAN specific API <cran.r-project.org/package=\%s&version=\%s> to install a package.
As this is an archived copy we only use source version.
}
\examples{
\dontrun{
cran_version("repo.data", "0.1.5", lib = tempdir())
}
}
\references{
CRAN pages.
}
