% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_utils.R
\name{duplicated_alias}
\alias{duplicated_alias}
\title{Report duplicated alias}
\usage{
duplicated_alias(alias)
}
\arguments{
\item{alias}{The output of \code{\link[=cran_alias]{cran_alias()}} or \code{\link[=base_alias]{base_alias()}}}
}
\value{
A sorted data.frame with the Target, Package and Source of the duplicate alias.
}
\description{
Report duplicated alias
}
\examples{
# Checking the overlap between to seemingly unrelated packages:
alias <- alias(c("fect", "gsynth"))
if (length(alias)) {
  dup_alias <- duplicated_alias(alias)
  head(dup_alias)
}
}
\seealso{
Other utilities:
\code{\link[=cran_date]{cran_date()}},
\code{\link[=cran_doom]{cran_doom()}},
\code{\link[=cran_snapshot]{cran_snapshot()}},
\code{\link[=package_date]{package_date()}},
\code{\link[=package_repos]{package_repos()}},
\code{\link[=repos_dependencies]{repos_dependencies()}},
\code{\link[=update_dependencies]{update_dependencies()}}
}
\concept{utilities}
