% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{os_alias}
\alias{os_alias}
\title{Base R OS specific alias}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 33 rows and 5 columns.
}
\usage{
os_alias
}
\value{
A matrix with 33 rows and 5 columns:
\describe{
\item{Package}{Package name}
\item{os}{OS system where this applies}
\item{file}{Name of the Rd file.}
\item{Source}{Path to the file.}
\item{Target}{Name of the Target}
}
}
\description{
Data from the R source code of OS specific man help pages.
This is to complement \code{tools::base_aliases_db()} which only provides links for Unix.
}
\examples{
os_alias

}
\keyword{datasets}
