% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_date.R
\name{package_date_actions}
\alias{package_date_actions}
\title{Package dates}
\usage{
package_date_actions(packages = ".", which = "strong")
}
\arguments{
\item{packages}{Name of the package on CRAN.
It accepts also local path to packages source directories but then the
function works as if the package is not released yet.}

\item{which}{a character vector listing the types of
    dependencies, a subset of
    \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.
    Character string \code{"all"} is shorthand for that vector,
    character string \code{"most"} for the same vector without
    \code{"Enhances"},
    character string \code{"strong"} (default) for the first three
    elements of that vector.
  }
}
\value{
A vector with the time when the package was publish and when were the dependencies available.
}
\description{
Same as \code{\link[=package_date]{package_date()}} but using CRAN's actions instead of public archive.
}
\details{
This provides information about when a package was removed or archived for a
more accurate estimation.
}
\examples{
package_date_actions("afmToolkit")
}
