% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{pkges_repos}
\alias{pkges_repos}
\title{Title}
\usage{
pkges_repos(repos = getOption("repos"))
}
\arguments{
\item{repos}{}
}
\value{
A data.frame with one line per package and at least one column per repository.
}
\description{
Title
}
\examples{
repos()
}
