% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrefs_utils.R
\name{targets2files}
\alias{targets2files}
\title{Resolves missing targets}
\usage{
targets2files(links, alias)
}
\arguments{
\item{links}{The output of \code{\link[=split_anchor]{split_anchor()}}.}

\item{alias}{The output of \code{\link[=alias2df]{alias2df()}} as data.frame.}
}
\value{
A data.frame with to_pkg, to_target, from_pkg, from_Rd, n, to_Rd.
}
\description{
Resolves links that require to know available alias so solve them.
}
\keyword{internal}
