% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{alias}
\alias{alias}
\title{Links}
\usage{
alias(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with three columns: Package, Source and Target.
NA if not able to collect the data from the repository.
}
\description{
Retrieve links of repository packages to other packages' documentation.
}
\note{
For completeness it also provides the alias of R packages themselves.
}
\examples{
oldrepos <- getOption("repos")
setRepositories(ind = c(1, 2), addURLs = "https://cran.r-project.org")
# show repositories
getOption("repos")
\donttest{
head(alias(c("ggplot2", "BiocCheck")))
}

# Clean  up
options(repos = oldrepos)
}
\seealso{
Other alias: 
\code{\link{base_alias}()},
\code{\link{cran_alias}()}
}
\concept{alias}
\concept{meta info}
