% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_help.R
\name{base_help_cliques}
\alias{base_help_cliques}
\title{Help pages with cliques}
\usage{
base_help_cliques()
}
\value{
Return a data.frame of help pages not connected to the network of help pages.
\code{NA} if not able to collect the data from CRAN.
}
\description{
Some help pages have links to and from but they are closed networks.
}
\details{
Requires igraph
}
\examples{
\donttest{
if (requireNamespace("igraph", quietly = TRUE)) {
  base_help_cliques()
}
}
}
\seealso{
Other functions related to BASE help pages:
\code{\link[=base_help_pages_not_linked]{base_help_pages_not_linked()}},
\code{\link[=base_help_pages_wo_links]{base_help_pages_wo_links()}}
}
\concept{functions related to BASE help pages}
