% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_help.R
\name{base_help_pages_not_linked}
\alias{base_help_pages_not_linked}
\title{Help pages without links}
\usage{
base_help_pages_not_linked()
}
\value{
A data.frame with two columns: Package and Source.
\code{NA} if not able to collect the data from CRAN.
}
\description{
Help pages without links to other help pages.
This makes harder to navigate to related help pages.
}
\examples{
\donttest{
bhnl <- base_help_pages_not_linked()
head(bhnl)
}
}
\seealso{
Other functions related to BASE help pages: 
\code{\link{base_help_cliques}()},
\code{\link{base_help_pages_wo_links}()}
}
\concept{functions related to BASE help pages}
