% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_help.R
\name{base_help_pages_wo_links}
\alias{base_help_pages_wo_links}
\title{Help pages not linked from base R}
\usage{
base_help_pages_wo_links()
}
\value{
A data.frame with two columns: Package and Source
}
\description{
Help pages without links from other help pages.
This makes harder to find them.
}
\examples{
\donttest{
bhwl <- base_help_pages_wo_links()
head(bhwl)
}
}
\seealso{
Other functions related to BASE help pages: 
\code{\link{base_help_cliques}()},
\code{\link{base_help_pages_not_linked}()}
}
\concept{functions related to BASE help pages}
