% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrefs_base.R
\name{base_links}
\alias{base_links}
\title{R's links}
\usage{
base_links()
}
\value{
A data.frame with the links on R's files.
It has 7 columns
}
\description{
Retrieve links on R documentation files.
}
\examples{
bl <- base_links()
head(bl)
}
