% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor.R
\name{bioc_cran_archived}
\alias{bioc_cran_archived}
\title{Bioconductor packages using CRAN archived packages}
\usage{
bioc_cran_archived(which = "strong")
}
\arguments{
\item{which}{a character vector listing the types of
    dependencies, a subset of
    \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.
    Character string \code{"all"} is shorthand for that vector,
    character string \code{"most"} for the same vector without
    \code{"Enhances"},
    character string \code{"strong"} (default) for the first three
    elements of that vector.
  }
}
\value{
A data.frame with the name of the Bioconductor packages
depending on archived packages (on Archived column) and the
number of missing packages (n).
\code{NA} if not able to collect the data.
}
\description{
Checks on the 4 Bioconductor repositories which packages depend on a
archived package.
}
\examples{
bca <- bioc_cran_archived()
head(bca)
}
\seealso{
For CRAN's data source: \code{\link[tools:CRAN_package_db]{tools::CRAN_package_db()}}
}
