% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor.R
\name{bioc_cran_archived}
\alias{bioc_cran_archived}
\title{Bioconductor packages using CRAN archived packages}
\usage{
bioc_cran_archived()
}
\value{
A character vector with the name of the Bioconductor packages
depending on CRAN packages that were archived.
}
\description{
Checks on the 4 Bioconductor repositories which packages depend on a CRAN
archived package.
}
\examples{
bioc_cran_archived()
}
