% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_actions.R
\name{cran_actions}
\alias{cran_actions}
\title{Look at the CRAN actions db}
\usage{
cran_actions(packages = NULL, silent = FALSE)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with Date, Time, User, Action, Package and Version columns.
\code{NA} if not able to collect the data from CRAN.
}
\description{
CRAN tracks movements of packages and the actions used (for example to report
the number of manual actions taken by the volunteers).
}
\keyword{internal}
