% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_actions.R
\name{cran_actions}
\alias{cran_actions}
\title{Look at the CRAN actions db}
\usage{
cran_actions(packages = NULL, silent = FALSE)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}

\item{silent}{A logical value to issue warnings about the data or not.}
}
\value{
A data.frame with Date, Time, User, Action, Package and Version columns.
\code{NA} if not able to collect the data from CRAN.
}
\description{
CRAN tracks movements of packages and the actions used (for example to report
the number of manual actions taken by the volunteers).
}
\details{
There are three possible actions with packages source code: publish, archive and remove.
\itemize{
\item Publish: Add it to CRAN's PACAKGES file, users can install that version.
\item Archive: Removed from CRAN's repository PACKAGES file so users can't access the package with \code{\link[=available.packages]{available.packages()}}.
Remains on CRAN archive: \url{https://cran.r-project.org/src/contrib/Archive/}.
\item Remove: Removed from CRAN's archive.
}
}
\examples{
ca <- cran_actions(silent = TRUE)
head(ca)
}
