% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_alias.R
\name{cran_alias}
\alias{cran_alias}
\title{CRAN's alias}
\usage{
cran_alias(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with three columns: Package, Source and Target.
\code{NA} if not able to collect the data from CRAN.
}
\description{
Retrieve alias available on CRAN.
}
\examples{
\donttest{
ca <- cran_alias("BWStest")
head(ca)
}
}
\seealso{
The raw source of the data is: \code{\link[tools:CRAN_aliases_db]{CRAN_aliases_db()}}.

Other alias:
\code{\link[=alias]{alias()}},
\code{\link[=base_alias]{base_alias()}}

Other meta info from CRAN:
\code{\link[=cran_actions]{cran_actions()}},
\code{\link[=cran_archive]{cran_archive()}},
\code{\link[=cran_comments]{cran_comments()}},
\code{\link[=cran_doom]{cran_doom()}},
\code{\link[=cran_links]{cran_links()}},
\code{\link[=links]{links()}}
}
\concept{alias}
\concept{meta info from CRAN}
