% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_alias.R
\name{cran_alias}
\alias{cran_alias}
\title{CRAN's alias}
\usage{
cran_alias()
}
\value{
A data.frame with three columns: Package, Source and Target.
}
\description{
Retrieve alias available on CRAN.
}
\examples{
ca <- cran_alias()
head(ca)
}
\seealso{
\code{\link[tools:CRANtools]{tools::CRAN_aliases_db()}}

Other alias: 
\code{\link{base_alias}()}
}
\concept{alias}
