% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_alias.R
\name{cran_alias}
\alias{cran_alias}
\title{CRAN's alias}
\usage{
cran_alias(package = NULL)
}
\arguments{
\item{package}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with three columns: Package, Source and Target.
}
\description{
Retrieve alias available on CRAN.
}
\examples{
ca <- cran_alias()
head(ca)
}
\seealso{
The raw source of the data is: \code{\link[tools:CRANtools]{tools::CRAN_aliases_db()}}

Other alias: 
\code{\link{base_alias}()}
}
\concept{alias}
