% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_archive.R
\name{cran_archive}
\alias{cran_archive}
\title{Retrieve CRAN archive}
\usage{
cran_archive(packages = NULL)
}
\value{
A data.frame with 6 columns: Package, Date (of publication), Version,
User, size and status (archived or current).
It is sorted by package name and date
}
\description{
Retrieve the archive and the current database.
}
\details{
There are some packages with NA in version, those are version names that do
not pass current \code{\link[=package_version]{package_version()}} with \code{strict = TRUE}.
Other packages might have been on CRAN but could have been removed.
}
\examples{
\donttest{
ca <- cran_archive()
head(ca)
}
}
\seealso{
\code{\link[=CRAN_archive_db]{CRAN_archive_db()}}, \code{\link[=CRAN_current_db]{CRAN_current_db()}}, \code{\link[=cran_comments]{cran_comments()}}.
}
