% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_archive.R
\name{cran_comments}
\alias{cran_comments}
\title{CRAN comments}
\usage{
cran_comments()
}
\value{
A data.frame with four columns: package, comment, date and action.
}
\description{
CRAN volunteers document since ~2009 why they archive packages.
This function retrieves the data and prepares it for its usage.
}
\details{
The comments are slightly edited: multiple comments for the same action that
they took are joined together.
Actions are inferred from keywords.
}
\examples{
\donttest{
cc <- cran_comments()
head(cc)
}
}
\references{
Original file: \url{https://cran.r-project.org/src/contrib/PACKAGES.in}
}
