% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_snapshot.R
\name{cran_date}
\alias{cran_date}
\alias{cran_session}
\title{Estimate CRAN's date of packages}
\usage{
cran_date(versions)

cran_session(session = sessionInfo())
}
\arguments{
\item{versions}{A data.frame with the packages names and versions}

\item{session}{Session information.}
}
\value{
Last installation date from CRAN.
}
\description{
Check which CRAN dates are possible for a given packages and versions.
}
\examples{
ip <- installed.packages()
cran_date(ip)
cran_session()
}
