% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_help.R
\name{cran_help_cliques}
\alias{cran_help_cliques}
\title{Help pages with cliques}
\usage{
cran_help_cliques(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
Return a data.frame of help pages not connected to the network of help pages.
Or NULL if nothing are found.
\code{NA} if not able to collect the data from CRAN.
}
\description{
Some help pages have links to other pages and they might be linked from others
but they are closed network: there is no link that leads to different help pages.
Each group of linked help pages is a clique.
}
\details{
The first clique is the biggest one.
You might want to check if others cliques can be connected to this one.

Requires igraph.
}
\examples{
\dontshow{if (requireNamespace("igraph", quietly = TRUE)) withAutoprint(\{ # examplesIf}
chc <- cran_help_cliques("BaseSet")
if (!is.null(dim(chc))) {
   table(chc$clique)
   chc[chc$clique != 1L, ]
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other functions related to CRAN help pages: 
\code{\link{cran_help_pages_not_linked}()},
\code{\link{cran_help_pages_wo_links}()}
}
\concept{functions related to CRAN help pages}
