% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_help.R
\name{cran_help_cliques}
\alias{cran_help_cliques}
\title{Help pages with cliques}
\usage{
cran_help_cliques(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
Return a data.frame of help pages not connected to the network of help pages.
}
\description{
Some help pages have links to and from but they are closed networks.
}
\details{
Requires igraph
}
\examples{
chc <- cran_help_cliques()
head(chc)
}
\seealso{
Other cran_help_pages: 
\code{\link{cran_help_pages_not_linked}()},
\code{\link{cran_help_pages_wo_links}()}
}
\concept{cran_help_pages}
