% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_help.R
\name{cran_help_pages_wo_links}
\alias{cran_help_pages_wo_links}
\title{Help pages not linked}
\usage{
cran_help_pages_wo_links(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with two columns: Package and Source
\code{NA} if not able to collect the data from CRAN.
}
\description{
Help pages without links from other help pages.
This makes harder to find them.
}
\examples{
\donttest{
ap <- available.packages()
if (NROW(ap)) {
    a_package <- rownames(ap)[startsWith(rownames(ap), "a")][1]
    chwl <- cran_help_pages_wo_links(a_package)
    head(chwl)
}
}
}
\seealso{
Other functions related to CRAN help pages: 
\code{\link{cran_help_cliques}()},
\code{\link{cran_help_pages_not_linked}()}
}
\concept{functions related to CRAN help pages}
