% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_help.R
\name{cran_help_pages_wo_links}
\alias{cran_help_pages_wo_links}
\title{Help pages not linked from base R.}
\usage{
cran_help_pages_wo_links()
}
\value{
A data.frame with two columns: Package and Source
}
\description{
Help pages without links from other help pages.
This makes harder to find them.
}
\examples{
chwl <- cran_help_pages_wo_links()
head(chwl)
}
\seealso{
Other cran_help_pages: 
\code{\link{cran_help_cliques}()},
\code{\link{cran_help_pages_not_linked}()}
}
\concept{cran_help_pages}
