% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_xrefs.R
\name{cran_links}
\alias{cran_links}
\title{CRAN's links}
\usage{
cran_links(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with the links on CRAN's packages.
It has 4 columns: Package, Anchor, Target and Source.
\code{NA} if not able to collect the data from CRAN.
}
\description{
Retrieve links on CRAN packages' R documentation files.
}
\examples{
cl <- cran_links("CytoSimplex")
head(cl)
}
\seealso{
The raw source of the data is: \code{\link[tools:CRAN_rdxrefs_db]{CRAN_rdxrefs_db()}}.

Other links from CRAN:
\code{\link[=cran_pages_links]{cran_pages_links()}},
\code{\link[=cran_pkges_links]{cran_pkges_links()}},
\code{\link[=cran_targets_links]{cran_targets_links()}},
\code{\link[=links]{links()}}

Other meta info from CRAN:
\code{\link[=cran_actions]{cran_actions()}},
\code{\link[=cran_alias]{cran_alias()}},
\code{\link[=cran_archive]{cran_archive()}},
\code{\link[=cran_comments]{cran_comments()}},
\code{\link[=cran_doom]{cran_doom()}},
\code{\link[=links]{links()}}
}
\concept{links from CRAN}
\concept{meta info from CRAN}
