% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrefs_cran.R
\name{cran_links}
\alias{cran_links}
\title{CRAN's links}
\usage{
cran_links()
}
\value{
A data.frame with the links on CRAN's packages
}
\description{
Retrieve links on CRAN packages' R documentation files.
}
\examples{
cl <- cran_links()
head(cl)
}
