% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_xrefs.R
\name{cran_links}
\alias{cran_links}
\title{CRAN's links}
\usage{
cran_links()
}
\value{
A data.frame with the links on CRAN's packages.
It has 4 columns: Package, Anchor, Target and Source.
}
\description{
Retrieve links on CRAN packages' R documentation files.
}
\examples{
cl <- cran_links()
head(cl)
}
\seealso{
\code{\link[tools:CRANtools]{tools::CRAN_rdxrefs_db()}}

Other links: 
\code{\link{base_links}()}
}
\concept{links}
