% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_xrefs.R
\name{cran_pages_links}
\alias{cran_pages_links}
\title{Links between help pages by page}
\usage{
cran_pages_links(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with 6 columns: from_pkg, from_Rd, to_pkg, to_Rd, n (Number of links).
}
\description{
Explore the relationship between CRAN packages and other help pages.
If the target help page is ambiguous it is omitted.
}
\examples{
cpl <- cran_pages_links("Matrix")
head(cpl)
}
\seealso{
Other links from CRAN: 
\code{\link{cran_links}()},
\code{\link{cran_pkges_links}()},
\code{\link{cran_targets_links}()}
}
\concept{links from CRAN}
