% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_snapshot.R
\name{cran_snapshot}
\alias{cran_snapshot}
\title{Check CRAN package state on any given date}
\usage{
cran_snapshot(date)
}
\arguments{
\item{date}{The date you want to check.}
}
\value{
The data.frame with the packages and versions at a given date.
\code{NA} if not able to collect the data from CRAN.
}
\description{
Given the available information which packages were on CRAN on a given date?
}
\note{
Due to missing of CRAN comments some packages are not annotated when
were they archived and more packages than present might be returned for any
given date.
}
\examples{
\donttest{
cs <- cran_snapshot(Sys.Date() -2 )
head(cs)
}
}
\seealso{
Other utilities:
\code{\link[=cran_date]{cran_date()}},
\code{\link[=cran_doom]{cran_doom()}},
\code{\link[=duplicated_alias]{duplicated_alias()}},
\code{\link[=package_date]{package_date()}},
\code{\link[=package_repos]{package_repos()}},
\code{\link[=repos_dependencies]{repos_dependencies()}},
\code{\link[=update_dependencies]{update_dependencies()}}
}
\concept{utilities}
