% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_snapshot.R
\name{cran_snapshot}
\alias{cran_snapshot}
\title{Check CRAN package state}
\usage{
cran_snapshot(date)
}
\arguments{
\item{date}{The date you want to check.}
}
\value{
The data.frame with the packages and versions at a given date.
}
\description{
Check CRAN package state
}
\examples{
cran_snapshot(Sys.Date() -2 )
}
