% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_snapshot.R
\name{cran_snapshot}
\alias{cran_snapshot}
\title{Check CRAN package state}
\usage{
cran_snapshot(date)
}
\arguments{
\item{date}{The date you want to check.}
}
\value{
The data.frame with the packages and versions at a given date.
}
\description{
Given the available information which packages were on CRAN on a given date?
}
\note{
Due to missing of CRAN comments some packages are not annotated when
were they archived and more packages will be returned for any given date.
}
\examples{
cs <- cran_snapshot(Sys.Date() -2 )
head(cs)
}
