% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{links}
\alias{links}
\title{Links}
\usage{
links(packages = NULL)
}
\arguments{
\item{packages}{A vector with packages or \code{NULL} for all packages.}
}
\value{
A data.frame with the links on packages.  It has 4 columns: Package, Anchor, Target and Source.
NA if not able to collect the data from the repository.
}
\description{
Retrieve links of repository packages to other packages' documentation.
}
\examples{
oldrepos <- getOption("repos")
setRepositories(ind = c(1, 2), addURLs = "https://cran.r-project.org")
head(links(c("ggplot2", "BiocCheck")))

# Clean  up
options(repos = oldrepos)
}
\seealso{
Other links from CRAN: 
\code{\link{cran_links}()},
\code{\link{cran_pages_links}()},
\code{\link{cran_pkges_links}()},
\code{\link{cran_targets_links}()}

Other meta info from CRAN: 
\code{\link{cran_actions}()},
\code{\link{cran_alias}()},
\code{\link{cran_archive}()},
\code{\link{cran_comments}()},
\code{\link{cran_doom}()},
\code{\link{cran_links}()}
}
\concept{links from CRAN}
\concept{meta info from CRAN}
