% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_date.R
\name{package_date}
\alias{package_date}
\title{Find earliest date of compatibility}
\usage{
package_date(pkg = ".", which = "all")
}
\arguments{
\item{pkg}{Path to the package folder or name of the package published.}

\item{which}{a character vector listing the types of
    dependencies, a subset of
    \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.
    Character string \code{"all"} is shorthand for that vector,
    character string \code{"most"} for the same vector without
    \code{"Enhances"},
    character string \code{"strong"} (default) for the first three
    elements of that vector.
  }
}
\value{
A vector with the datetimes of the published package (or current
time if not published) and the date the required restrictions are met.
}
\description{
Search the DESCRIPTION file for the release dates of dependencies and return the earliest date.
This is the date at which the package could be installed.
}
\details{
Currently this function assumes that packages only use ">=" and not other operators.
This might change on the future if other operators are more used.
}
\examples{
package_date("ABACUS")
}
