% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_date.R
\name{package_date_actions}
\alias{package_date_actions}
\title{Package dates}
\usage{
package_date_actions(pkg = ".", which = "strong")
}
\description{
Same as package_date but using CRAN's actions instead of public archive.
}
\details{
This provides information about when a package was removed or archived for a
more accurate estimation.
}
\examples{
package_date_actions("afmToolkit")
}
\keyword{internal}
