% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{package_dependencies}
\alias{package_dependencies}
\title{Find current installations}
\usage{
package_dependencies(pkg = ".", which = "strong")
}
\arguments{
\item{pkg}{Path to a file with a DESCRIPTION file or a the name of a package.}

\item{which}{a character vector listing the types of
    dependencies, a subset of
    \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.
    Character string \code{"all"} is shorthand for that vector,
    character string \code{"most"} for the same vector without
    \code{"Enhances"},
    character string \code{"strong"} (default) for the first three
    elements of that vector.
  }
}
\value{
A data.frame with the name, version required and version used.
}
\description{
Despite the description minimal requirements find which versions are
required due to dependencies.
}
\examples{
pd <- package_dependencies("ggeasy")
head(pd)
}
