% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.data-package.R
\docType{package}
\encoding{UTF-8}
\name{repo.data-package}
\alias{repo.data}
\alias{repo.data-package}
\title{repo.data: R Repository Data}
\description{
Retrieve metadata about packages from repositories to explore package dependencies, links between help pages, aliases, package availability on a given date, and other repository dependent outcome. This metadata can be used to help package maintainers and users to navigate changes on dependencies and with reproducibility.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/llrs/repo.data}
  \item \url{https://llrs.github.io/repo.data/}
  \item Report bugs at \url{https://github.com/llrs/repo.data/issues}
}

}
\author{
\strong{Maintainer}: Lluís Revilla Sancho \email{lluis.revilla@gmail.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID}) [copyright holder]

}
\keyword{internal}
