% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.data-package.R
\docType{package}
\encoding{UTF-8}
\name{repo.data-package}
\alias{repo.data}
\alias{repo.data-package}
\title{repo.data: Get repository data}
\description{
Retrieve CRAN's data and prepare repository data like the one provided by it.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/llrs/repo.data}
  \item Report bugs at \url{https://github.com/llrs/repo.data/issues}
}

}
\author{
\strong{Maintainer}: Lluís Revilla Sancho \email{lluis.revilla@gmail.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID})

}
\keyword{internal}
