% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{repos_dependencies}
\alias{repos_dependencies}
\title{Tidy dependencies}
\usage{
repos_dependencies(which = "all")
}
\value{
A data.frame with 5 columns: the name of the dependency,
the operator (op), the version it depends the type of dependency and the package.
}
\description{
Extract the packages dependencies, name of the dependency, operator and version
for each type and package of current repositories (\code{getOptions("repos")}).
}
\examples{
rd <- repos_dependencies()
head(rd)
}
