% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{update_dependencies}
\alias{update_dependencies}
\title{Upgradable versions}
\usage{
update_dependencies(packages)
}
\arguments{
\item{packages}{A character vector of packages names.}
}
\value{
The data.frame filtered with some relevant rows.
\code{NA} if not able to collect the data from repositories.
}
\description{
Helper function to detect which package have a required version on the
dependencies that could be upgraded.
}
\details{
Increasing this version requirements won't affect users as they already
should have these versions installed as required by other dependencies.
}
\examples{
update_dependencies("arrow")
}
\seealso{
\code{\link[=package_dependencies]{package_dependencies()}}

Other utilities: 
\code{\link{cran_date}()},
\code{\link{cran_doom}()},
\code{\link{cran_snapshot}()},
\code{\link{duplicated_alias}()},
\code{\link{package_date}()},
\code{\link{package_repos}()},
\code{\link{repos_dependencies}()}
}
\concept{utilities}
