% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{update_dependencies}
\alias{update_dependencies}
\title{Check versions}
\usage{
update_dependencies(deps)
}
\arguments{
\item{deps}{The output of \code{\link[=package_dependencies]{package_dependencies()}} or any data.frame with a
version and required column.}
}
\value{
The data.frame filtered with some relevant rows
}
\description{
Helper function to detect which package have a required version on the
dependencies that could be upgraded.
}
\examples{
pd <- package_dependencies("arrow")
update_dependencies(pd)
}
\seealso{
\code{\link[=package_dependencies]{package_dependencies()}}

Other utilities: 
\code{\link{cran_date}()},
\code{\link{cran_doom}()},
\code{\link{cran_snapshot}()},
\code{\link{package_date}()},
\code{\link{pkges_repos}()},
\code{\link{repos_dependencies}()}
}
\concept{utilities}
